move ktb-edit-service-file-list to _components
