+++ bridge/client/app/_components/ktb-edit-service-file-list/ktb-edit-service-file-list.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { KtbEditServiceFileListComponent } from './ktb-edit-service-file-list.component';
-import { AppModule } from '../app.module';
+import { AppModule } from '../../app.module';
 import { By } from '@angular/platform-browser';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
